Ensure space for NULL byte
Despite that we ultimately want to drop the NULL byte, we need to
include space for it in the buffer provided to libc's vsnprintf;
otherwise it may replace the last byte of the string with a NULL byte.

This also simplifies things by providing a zero-length buffer for the
first call to vsnprintf, and then allocating exactly the required space.
This gives only one code path to maintain (as opposed to 2-3 before
depending if the initial buffer was less than, equal to, or greater than
the initial size). While this means paying to format the string twice,
it lets us drop `shrink_to_fit` of the final string, which can be a
relatively expensive call to the allocator.

Fixes https://github.com/shadow/shadow/issues/1432