Be robust to changing buffer length requirement
vsnprintf *should* return the exact number of characters that will be
printed with the given arguments. In Shadow, I've seen at least one
instance of it wanting more space on the second call, though. This could
be a bug in Shadow (maybe another thread mutating one of the arguments
in between calls, but is difficult to debug.

Probably better to be robust here and just reallocate if needed than to
crash.